graphics.h, graphics.cpp, model.cpp: change rounding semantics to make things easier for the optimizer.

+++ src/graphics.cpp
@@ -210,7 +210,7 @@ bool initialize_graphics (program_t & program)
 }
 
 void program_t::set_view (const float (& view) [4],
-                          float width, float height,
+                          int width, int height,
                           float fog_near, float fog_far,
                           float line_width_extra, float line_sharpness)
 {
